Reorganise unit tests
